04_SQL kunskapskontroll fil
Here can you find all the necessary information for the kunskapskontroll. All of the answers are formally answered at the end of the analysis, but some of it is also written throughout the code as I explored and took notes in order to complete the descriptive statistics of the dataset. I hope this is okay, and if you have any questions about it then let me know.

Mvh 

Michael